add shillSet example
Let's try and make the shill example from the BRTs paper work. I'm going
to start with a set for each capability and see if we can get that to
work before I add a map of sets to keep them together.

After we get this example working, we should get RJ's state example
working before we mapify everything.