isolated tabs/ifun bug and split into own test
The problem is that when we have an ~> under an forall, at the term
level we have a tabs, and that means we have to swtich from checking (at
the type with the implicit parameter) to synthing (plus a subtyping
check). But when we synth, this implicit parameter is simply not in
scope: since it's not bound anywhere in the term, it can't be added to
the context during synthesis.